tests/wpt/web-platform-tests/css/css-animations/parsing/animation-delay-end-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-delay-end">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value("animation-delay-end", "initial", "0s");
test_computed_value("animation-delay-end", "-500ms", "-0.5s");
test_computed_value("animation-delay-end", "calc(2 * 3s)", "6s");
test_computed_value("animation-delay-end", "20s, 10s");
</script>
